view: handle view menu
NautilusToolbar handles the view menu, requiring
direct access to the underlying view.

Since we're wiping out every access to the underlying
view, this code is a hazard.

To fix that, makes the view handle the view widget.
Since we're making NautilusWindowSlot a wrapper, add
the necessary properties for it to expose view data
without exposing the view itself.